refactor: test contract functionality in pop_contracts
