+++ crates/pop-contracts/src/lib.rs
@@ -1,5 +1,7 @@
 mod build;
 mod new;
+mod test;
 
 pub use build::build_smart_contract;
 pub use new::create_smart_contract;
+pub use test::{test_e2e_smart_contract, test_smart_contract};
